Use a method for getting the main editor and projectmanager instance
For unit testing the plan is to not do any mocking for the most, but
simply load all dependencies without instantiating the main editor instance.
This will make sure there are no side effects, which would cause the editor
to get fully loaded, which isn't possible without jsdom and all sorts of other
mocks.
So every unit test will still load all dependencies, but only the import logic.
As long as no modules have any side effects, this shouldn't be a problem.
Side effects in modules should be avoided anyway, so if side effects cause
issues in the tests, that would be a good thing. The failing tests would
expose the anti-pattern.