feat(gizmos): Use entities for draggable translations
Instead of using a single position vector for draggables,
we'll use entities. This way they can be added to a hierarchy
and benefit from translations from parent benefits.
This also implements the change in the TranslationGizmo, this way
the draggables are always on the correct position when moving
the TranslationGizmo around