test(assets): Add first part of embedded asset creation e2e test
This is the first part of the embedded asset e2e test. The test
only creates an embedded asset for now, but doesn't actually reload
the page and check if loading works yet. This is because in order to
do this we would need to have a way to wait for the indexeddb filesystem
to finish writing.