fix(binary serialization): Encode data as reference ids for all structure references
Previously, we would only assign a reference ids to objects that were reoccurring,
but there was no way for the deserializer to know if an object was encoded as
reference id or inline. This change makes it so that if only a single object is
encoded as reference id, any objects that are reoccurring within the structure
are encoded as reference id as well.