fix(binary serialization): Don't enforce reference ids for variable length arrays
Initially we were always using reference ids for variable length arrays,
because there was no good support for recursively calling `parseStructureDigestable()`.
But now that we're calling it recursively for union types, we can use it for
variable length arrays as well. This way there is also no longer the need
for digestables with a "structureRef" in two separate places, as this is now
taken care of when calling `parseStructureDigestable` one level below.