refactor(binary serialization): Perform deserialization in a single pass
Previously we were creating digestables from the structure first, and
then in a second pass we would loop over all the digestables and parse
the binary. However, performing this in two passes was unnecessary and
made the code difficult to work with. With this new approach the structure
is traversed and parsed at the same time in a single pass. This makes it
more clear what data is being used and how. And as a nice bonus this should
make memory usage lower, since there is no longer a long flattened
digestables array. In the case of union arrays, the flattened array contained
all available union types, which might be unnecessary if the binary data
was only using a single type of the union. With the new single pass system
union arrays are only parsed as they are needed.