fix(droppablegui): Fix error when dropping on a gui which supports all asset types (#35)
The `supportedAssetTypes` option of droppable guis initially required `ProjectAsset` to be provided in order to constrain droppable guis to only assets. But dropping anything other than assets wasn't actually working at all.
This change removes this requirement and instead dropping is always only constrained to assets.

Lack of typing also caused an error to be thrown in cases where the `supportedAssetTypes` option was an empty array. This should be fixed as well.