feat(30): Better handling of file system permissions (#51)
This adds an arrow to the root entry that can be used for requesting the permission prompt. If permission is denied the root treview collapses again and gets unselected.

This also adds some logic to automatically expand the root treeview if permission is granted. Some e2e tests depend on this treeview being expanded, and previously the root treeview was always expanded.