fix: Use a single connection for IndexedDbUtil (#136)
We were opening a new connection for every db request. This change makes it open only a single one and reuse that for every transaction.
When the connection gets garbage collected browsers should automatically close it as far as I can tell.
This change also ensures that the connection is closed when a database is deleted.