feat: Don't change selection when dragging assets (#163)
Calling focus() on an element seems to prevent dragstart events. So this change ensures treeviews are only focused when they don't have focus yet. Dragging an element also causes that element to get focus, so calling focus again isn't necessary.

However, this focus event causes the current selection to change, and as a result the properties window content is changed. To prevent this, we revert the focus to the previous element the moment a treeview is dragged. This causes a brief flicker while the focus is changed back and forth. Unfortunately there's little that can be done about this in regards to :focus-within css selectors. But the flickering of the content of the properties window could potentially be fixed in a follow up PR.