feat: Support for default colors in material map texture properties (#238)
This adds support for setting colors as default values in material maps. The WebGpuRenderer creates a temporary 1x1 texture with the provided color. These textures are automatically destroyed when the destructor of the material is called or once the material is garbage collected.

There are still some optimizations that can be made. For instance, the pipeline could be replaced with one that doesn't use textures at all and turns the color into a uniform. However, I think the main benefit of that would be that it allows you to frequently change the value of the color. Since material maps don't really have support for changing default values at runtime, I don't think this is really needed.
I'm not sure if there's any difference between binding a texture and using the existing uniform buffer. If so, using the uniform buffer might still be beneficial, but this would have to be tested first to see if that would be worth it.