feat: Add blend state to WebGpuPipelineConfig (#507)
This change removes fragment target formats from pipeline configs, they
are now controlled by the domtarget directly.
If we ever want to make the format configurable it should be done
through the domtarget, not the pipelineconfig.

Additionally support is added for adding blendstate configuration to
pipeline configs. The WebGpuRenderer then merges the outputFromat of the
domtarget with the blendstate of the pipelineconfig when creating the
pipeline.

This also adds a test for serialization and deserialization of
WebGpuPipelineConfigs with a function that can easily be reused by tests
for other asset types.

Finally, the preloadVertexStates property has been removed, it wasn't
used anyway so it would only confuse users.