feat: Don't resolve SingleInstancePromise.waitForFinishIfRunning when the queue is not empty (#530)
Fixes #529

Previously `waitForFinishIfRunning` would resolve as soon as the first
run was done. But more often than not when this function is used, you
want it to resolve after *all* runs are done.

At the moment it seems like there are no cases where you'd want the old
behaviour. If we ever need this again in the future we can either add a
new method or add a parameter to this method.