chore: Some PR deployment fixes
Disables auto merge, we want the preview to show the latest changes.
If this causes the deployment to fail then the PR submitter should merge
the main branch back in their PR.

Removes auto_inactive and task, not sure why but deployments are still
getting marked as inactive even with these values set. It seems like
it was working fine before this was added so we'll just revert it and
see what happens.